Spells/SpellClasses: De/Buff categorisation
