+++ Source/NexusForever.Shared/Network/Message/GameMessageOpcode.cs
@@ -228,9 +228,11 @@ namespace NexusForever.Shared.Network.Message
         ClientCancelEffect              = 0x0802,
         ServerCooldown                  = 0x0804,
         Server0811                      = 0x0811, // spell related: broadcast parts of 0x07FF?
+        ServerSpellBuffRemove           = 0x0813,
         Server0814                      = 0x0814, // spell related
         Server0816                      = 0x0816, // spell related: broadcast parts of 0x07FF?
         Server0817                      = 0x0817, // spell related
+        Server0818                      = 0x0818,
         ClientStorefrontPurchaseAccount = 0x0828,
         ClientStorefrontPurchaseCharacter = 0x082A,
         ClientStorefrontRequestCatalog  = 0x082D,
