+++ Source/NexusForever.WorldServer/Game/Spell/Static/SpellClass.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace NexusForever.WorldServer.Game.Spell.Static
+{
+    public enum SpellClass
+    {
+        BuffNonDispelRightClickOk = 0x000E,
+        BuffDispellable           = 0x0024,
+        BuffNonDispellable        = 0x0025,
+        DebuffDispellable         = 0x0026,
+        DebuffNonDispellable      = 0x0027
+    }
+}
