+++ Source/NexusForever.WorldServer/Network/Message/Model/ServerSpellBuffRemove.cs
@@ -1,0 +1,19 @@
+using System.Collections.Generic;
+using NexusForever.Shared.Network;
+using NexusForever.Shared.Network.Message;
+
+namespace NexusForever.WorldServer.Network.Message.Model
+{
+    [Message(GameMessageOpcode.ServerSpellBuffRemove)]
+    public class ServerSpellBuffRemove : IWritable
+    {
+        public uint CastingId { get; set; }
+        public uint CasterId { get; set; }
+
+        public void Write(GamePacketWriter writer)
+        {
+            writer.Write(CastingId);
+            writer.Write(CasterId);
+        }
+    }
+}
